seperate param_func file
