+++ select_timestep_allbhdata.py
@@ -1,41 +1,7 @@
 import numpy as np
 import multiprocessing
 from KWBHS import *
-
-
-PATH_ZOO = '/physics2/kuanweih/project_BH_seedmass/simulation_zoo/'
-PATH_RUNS = ['run_10Mpc_362341/Run_seed5e5/Con_2/']    # list of runs
-# TODO should I use list comprehension here?
-# TODO PATH_RUN[:16] will give 'run_10Mpc_362341'
-# TODO PATH_RUN[17:-7] will give 'Run_seed5e5'
-# TODO PATH_RUN[25:-7] will give '5e5'
-
-
-def npsaves(dir_name, redshifts, bhmasss, bhids, bhaccs,
-            bhrhos, bhcss, bhvels, bhxs, bhys, bhzs):
-    """ save bh data """
-    np.save('{}redshift'.format(dir_name), redshifts)
-    np.save('{}bhmass'.format(dir_name), bhmasss)
-    np.save('{}bhid'.format(dir_name), bhids)
-    np.save('{}bhacc'.format(dir_name), bhaccs)
-    np.save('{}bhrho'.format(dir_name), bhrhos)
-    np.save('{}bhcs'.format(dir_name), bhcss)
-    np.save('{}bhvel'.format(dir_name), bhvels)
-    np.save('{}bhx'.format(dir_name), bhxs)
-    np.save('{}bhy'.format(dir_name), bhys)
-    np.save('{}bhz'.format(dir_name), bhzs)
-
-
-def create_dir(dir_name):
-    """ create directory name according to the run """
-    import os
-    import errno
-    if not os.path.exists(os.path.dirname(dir_name)):
-        try:
-            os.makedirs(os.path.dirname(dir_name))
-        except OSError as exc:  # Guard against race condition
-            if exc.errno != errno.EEXIST:
-                raise
+from param_func import *
 
 
 def allbh_from_txt(path_run):
@@ -57,7 +23,8 @@ def allbh_from_txt(path_run):
     bhzs = kwbh.get_bhpos(2)
 
     # create directory and output data
-    dir_name = 'allbhdata/{}_{}/'.format(path_run[:16], path_run[25:-7])
+    dir_name = 'allbhdata/{}_{}/'.format(path_run[IDX_P1_I:IDX_P1_F],
+                                         path_run[IDX_P2_I:IDX_P2_F])
     create_dir(dir_name)
     npsaves(dir_name, redshifts, bhmasss, bhids, bhaccs,
             bhrhos, bhcss, bhvels, bhxs, bhys, bhzs)
