+++ select_timestep_mergerdata.py
@@ -1,45 +1,15 @@
 import numpy as np
 import multiprocessing
-
-
-PATH_ZOO = '/physics2/kuanweih/project_BH_seedmass/simulation_zoo/'
-PATH_RUNS = ['run_10Mpc_362341/Run_seed5e5/Con_2/']    # list of runs
-# TODO should I use list comprehension here?
-# TODO PATH_RUN[:16] will give 'run_10Mpc_362341'
-# TODO PATH_RUN[25:-7] will give '5e5'
-
-
-def npsaves(dir_name, redshifts, bhmasss, bhids, bhaccs,
-            bhrhos, bhcss, bhvels, bhxs, bhys, bhzs):
-    """ save bh data """
-    np.save('{}redshift'.format(dir_name), redshifts)
-    np.save('{}bhmass'.format(dir_name), bhmasss)
-    np.save('{}bhid'.format(dir_name), bhids)
-    np.save('{}bhacc'.format(dir_name), bhaccs)
-    np.save('{}bhrho'.format(dir_name), bhrhos)
-    np.save('{}bhcs'.format(dir_name), bhcss)
-    np.save('{}bhvel'.format(dir_name), bhvels)
-    np.save('{}bhx'.format(dir_name), bhxs)
-    np.save('{}bhy'.format(dir_name), bhys)
-    np.save('{}bhz'.format(dir_name), bhzs)
-
-
-def create_dir(dir_name):
-    """ create directory name according to the run """
-    import os
-    import errno
-    if not os.path.exists(os.path.dirname(dir_name)):
-        try:
-            os.makedirs(os.path.dirname(dir_name))
-        except OSError as exc:  # Guard against race condition
-            if exc.errno != errno.EEXIST:
-                raise
+from param_func import *
 
 
 def mergers_from_txt(path_run):
+    """ main function: find merger tree for the most massive BHs """
 
-    mmbh_dir = 'mmbhdata/{}_{}/'.format(path_run[:16], path_run[25:-7])
-    allbh_dir = 'allbhdata/{}_{}/'.format(path_run[:16], path_run[25:-7])
+    mmbh_dir = 'mmbhdata/{}_{}/'.format(path_run[IDX_P1_I:IDX_P1_F],
+                                        path_run[IDX_P2_I:IDX_P2_F])
+    allbh_dir = 'allbhdata/{}_{}/'.format(path_run[IDX_P1_I:IDX_P1_F],
+                                          path_run[IDX_P2_I:IDX_P2_F])
 
     # get merger id from mmbh ids
     mmbhids = np.load('{}bhid.npy'.format(mmbh_dir))
@@ -57,11 +27,11 @@ def mergers_from_txt(path_run):
     bhys = np.load('{}bhy.npy'.format(allbh_dir))
     bhzs = np.load('{}bhz.npy'.format(allbh_dir))
 
-    def by_merger_id(mgids, bhids, quantity):
-        con = np.array([False] * len(bhids))
-        for v in range(len(mgids)):
-            con = con + (bhids == mgids[v])
-        return quantity[con]
+    # def by_merger_id(mgids, bhids, quantity):
+    #     con = np.array([False] * len(bhids))
+    #     for v in range(len(mgids)):
+    #         con = con + (bhids == mgids[v])
+    #     return quantity[con]
 
     # merger bhid survive
     redshift = by_merger_id(mgids, bhids, redshifts)
@@ -76,10 +46,11 @@ def mergers_from_txt(path_run):
     bhz = by_merger_id(mgids, bhids, bhzs)
 
     # create directory and output data
-    dir_name = 'mergerdata/{}_{}/'.format(path_run[:16], path_run[25:-7])
+    dir_name = 'mergerdata/{}_{}/'.format(path_run[IDX_P1_I:IDX_P1_F],
+                                          path_run[IDX_P2_I:IDX_P2_F])
     create_dir(dir_name)
-    npsaves(dir_name, redshifts, bhmasss, bhids, bhaccs,
-            bhrhos, bhcss, bhvels, bhxs, bhys, bhzs)
+    npsaves(dir_name, redshift, bhmass, bhid, bhacc,
+            bhrho, bhcs, bhvel, bhx, bhy, bhz)
 
 
 p = multiprocessing.Pool(16)
